Use uint32_t as size

+++ model/palette.hpp
@@ -25,16 +25,16 @@ namespace Texpainter::Model
 			explicit Palette(DataBlock<Pixel>&& block): m_block{std::move(block)}
 			{}
 
-			explicit Palette(size_t n, Pixel color_init = black()):m_block{n}
+			explicit Palette(uint32_t n, Pixel color_init = black()):m_block{n}
 			{ std::fill(std::begin(m_block), std::end(m_block), color_init);}
 
-			auto operator[](size_t index) const
+			auto operator[](uint32_t index) const
 			{ return *(begin() + index); }
 
-			auto& operator[](size_t index)
+			auto& operator[](uint32_t index)
 			{ return *(std::begin(m_block) + index); }
 
-			void resize(size_t size_new)
+			void resize(uint32_t size_new)
 			{
 				auto offset_end = size();
 				Texpainter::resize(m_block, size_new);
@@ -48,7 +48,7 @@ namespace Texpainter::Model
 			DataBlock<Pixel> m_block;
 	};
 
-	inline void resize(Palette& p, size_t size_new)
+	inline void resize(Palette& p, uint32_t size_new)
 	{
 		p.resize(size_new);
 	}
